pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrgen"
version = "1.0.0"
description = "Synthetic text image generator for OCR training datasets"
authors = [{name = "SunjiwBorg"}]
dependencies = [
    "Pillow>=10.0.0",
    "numpy>=1.20.0",
]
requires-python = ">=3.8"

[project.scripts]
ocrgen = "ocrgen.main:main"